+++ test/session_test/single_access_token_test.rb
@@ -13,6 +13,14 @@ module SessionTest
         assert_equal "user_credentials", UserSession.params_key
       end
 
+      def test_headers_key
+        UserSession.headers_key = "my_headers_key"
+        assert_equal "my_headers_key", UserSession.headers_key
+
+        UserSession.headers_key "user_credentials"
+        assert_equal "user_credentials", UserSession.headers_key
+      end
+
       def test_single_access_allowed_request_types
         UserSession.single_access_allowed_request_types = ["my request type"]
         assert_equal ["my request type"], UserSession.single_access_allowed_request_types
@@ -28,11 +36,19 @@ module SessionTest
 
     class InstanceMethodsTest < ActiveSupport::TestCase
       def test_persist_persist_by_params
+        persist_persist_by(:params)
+      end
+
+      def test_persist_persist_by_headers
+        persist_persist_by(:headers)
+      end
+
+      def persist_persist_by(headers_or_params)
         ben = users(:ben)
         session = UserSession.new
 
         refute session.persisting?
-        set_params_for(ben)
+        send("set_#{headers_or_params}_for", ben)
 
         refute session.persisting?
         refute session.unauthorized_record
